updated jackson-databind to 2.9.7 to work around https://nvd.nist.gov/vuln/detail/cve-2018-7489.
Vulnerability in a nutshell: jackson-databind before 2.8.11.1 and 2.9.x before 2.9.5 allows unauthenticated remote code execution because of an incomplete fix for the cve-2017-7525 deserialization flaw. this is exploitable by sending maliciously crafted json input to the readvalue method of the objectmapper, bypassing a blacklist that is ineffective if the c3p0 libraries are available in the classpath.